// types/navigation.ts
export type RootStackParamList = {
    Home: undefined;
    Camera: undefined;
    File: undefined;
    Favourite: undefined;
    Gallery: { imageUri: string };  // Define the expected parameter for Gallery screen
    Recent: undefined;
    Generate: undefined;
    PdfViewer: { uri: string };
  //  Purchase: undefined;
    //Stripe: undefined;

  };
  